requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
anthropic>=0.25.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
cryptography>=42.0.0
pdfplumber>=0.11.0
apscheduler>=3.10.4
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
